requirements.txt
nvalues==1.0.0b6
pillow~=9.3
